services/dummy/requirements.txt
googleapis-common-protos
grpcio
grpcio-tools
grpcio-reflection
